Separated emojiRatings into separate component/test files
